fix a problem of max_int value

+++ src/index.c
@@ -6,7 +6,7 @@
 /*   By: massrayb <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/02/23 19:30:56 by massrayb          #+#    #+#             */
-/*   Updated: 2025/03/09 15:29:36 by massrayb         ###   ########.fr       */
+/*   Updated: 2025/03/12 17:10:15 by massrayb         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,16 +14,16 @@
 
 static t_node	*get_smallest_value(t_stack *stack_a)
 {
-	t_node			*node;
-	t_node			*smallest;
-	unsigned int	s_value;
+	t_node	*node;
+	t_node	*smallest;
+	long	s_value;
 
 	smallest = stack_a->top;
-	s_value = UINT_MAX;
+	s_value = LONG_MAX;
 	node = stack_a->top;
 	while (node)
 	{
-		if (((unsigned int)node->value < s_value && node->index == -1))
+		if (((long)node->value < s_value && node->index == -1))
 		{
 			smallest = node;
 			s_value = node->value;
